suddenly Tibetan was not displaying properly because //TIBETAN-SPECIFIC lines were cropping up in QD.java
yutso, please be careful: i think you copied QD.java from
Fieldling/build/src/unicode/ into your Fieldling/src/ directory.
this problem was a bit tricky for me to track down.

the bottom line is that you should never edit a source file
that contains //TIBETAN-SPECIFIC or //UNICODE-SPECIFIC
prefixes in it. only work with it if you see @TIBETAN@ and
@UNICODE@ prefixes.